COREX-611 - LLM Suggestions
Added EndpointConverter to JsonSerializerSettings to customize serialization of Sysmte.Net.IPEndPoint
