--- src/Logging/Loggly/Loggly/JsonSettings.cs
@@ -1,19 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace EMG.Extensions.Logging.Loggly
-{
-    public static class JsonSettings
-    {
-        public static readonly JsonSerializerSettings SerializerSettings = new JsonSerializerSettings
-        {
-            TypeNameHandling = TypeNameHandling.None,
-            NullValueHandling = NullValueHandling.Ignore,
-            Formatting = Formatting.None,
-            DefaultValueHandling = DefaultValueHandling.Include,
-            DateTimeZoneHandling = DateTimeZoneHandling.Utc,
-            DateFormatHandling = DateFormatHandling.IsoDateFormat,
-            ReferenceLoopHandling = ReferenceLoopHandling.Ignore,
-            Converters = { new FormattedIdConverter() }
-        };
-    }
-}
