+++ src/Logging/Loggly/Loggly/SerializerSettings/JsonSettings.cs
@@ -1,0 +1,18 @@
+﻿using Newtonsoft.Json;
+
+namespace EMG.Extensions.Logging.Loggly.SerializerSettings;
+
+public static class JsonSettings
+{
+    public static readonly JsonSerializerSettings SerializerSettings = new JsonSerializerSettings
+    {
+        TypeNameHandling = TypeNameHandling.None,
+        NullValueHandling = NullValueHandling.Ignore,
+        Formatting = Formatting.None,
+        DefaultValueHandling = DefaultValueHandling.Include,
+        DateTimeZoneHandling = DateTimeZoneHandling.Utc,
+        DateFormatHandling = DateFormatHandling.IsoDateFormat,
+        ReferenceLoopHandling = ReferenceLoopHandling.Ignore,
+        Converters = { new FormattedIdConverter(), new EndpointConverter() }
+    };
+}
